router/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
python-dotenv
psycopg2-binary
python-jose
pandas
python-multipart
httpx
asyncpg